Improve request logging
Add `ipHeader` option for `Stok#createServer()` to handle getting the client IP
when a proxy server is used.

Moves the request logger form `onPostHandler` to `onPreResponse` to handle
responses from `onPreHandler`.

Adds support for getting the status code form Boom error replies.

Drops dependency on `hapi-common-log` as it seems like it may no longer be
maintained and we needed to apply an update to handle Boom error replies.

Closes gh-14